+++ src/main/java/net/zhaiji/chestcavitybeyond/mixin/PlayerMixin.java
@@ -22,7 +22,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Player.class)
 public abstract class PlayerMixin extends LivingEntity {
-
     @Shadow
     protected FoodData foodData;